requirements.txt
# CartWise Pro - Python Dependencies
# ====================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP Requests
requests==2.31.0

# Serial Communication (RS485)
pyserial==3.5

# Environment Variables
python-dotenv==1.0.0

# Development Tools (Optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
